Refactored window names
